pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-forge"
version = "0.1.0"
description = "A sandboxed AI coding agent runtime using the ReAct pattern"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Aboubakar Koïta" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "click>=8.0",
    "fastapi>=0.115",
    "google-cloud-storage>=2.16",
    "httpx>=0.27",
    "docker>=7.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "structlog>=24.0",
    "uvicorn>=0.34",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
redis = [
    "redis>=5.0",
    "hiredis>=3.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "respx>=0.21",
    "types-docker>=7.0",
    "types-PyYAML>=6.0",
    "python-semantic-release>=9.0",
    "vulture>=2.11",
    "radon>=6.0",
]

[project.scripts]
agent-forge = "agent_forge.cli:main"

[project.urls]
Homepage = "https://github.com/akoita/agent-forge"
Repository = "https://github.com/akoita/agent-forge"
Issues = "https://github.com/akoita/agent-forge/issues"

[tool.hatch.build.targets.wheel]
packages = ["agent_forge"]

[tool.hatch.build.targets.wheel.force-include]
"agent_forge/profiles/builtins" = "agent_forge/profiles/builtins"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "C901", # McCabe complexity
    "S",    # flake8-bandit (security)
    "PT",   # flake8-pytest-style
    "RET",  # flake8-return
    "ARG",  # flake8-unused-arguments
    "ERA",  # flake8-eradicate (commented-out code)
    "PIE",  # flake8-pie (misc)
    "RUF",  # Ruff-specific rules
    "BLE",  # flake8-blind-except
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S108", "ARG001", "ARG002", "PT018", "RUF015"]
"tests/fixtures/**" = ["S201"]  # sample Flask app uses debug=True by design

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "redis.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring Docker (deselect with '-m \"not integration\"')",
    "e2e: marks end-to-end tests (deselect with '-m \"not e2e\"')",
]

# ---------------------------------------------------------------------------
# Semantic Release
# ---------------------------------------------------------------------------

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
branch = "main"
commit_message = "chore(release): v{version}"
build_command = ""
tag_format = "v{version}"

[tool.semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"

[tool.semantic_release.remote]
type = "github"
token = { env = "GH_TOKEN" }

[tool.semantic_release.publish]
upload_to_vcs_release = true
